#27 refactored controller directories
